remove the need to declare a GraphViewInternalOps when proxying one
